Cropper + DateTime Picker

+++ database/migrations/2024_01_01_181532_create_partners_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->double('lat');
             $table->double('lng');
             $table->enum('class',['أ','ب','ج','د']);
-            $table->integer('CR');
+            $table->BigInteger('CR');
             // logo to table file 
             $table->timestamps();
         });
